gn: Do stop sending NOTIFYs after receiving 477.
477 generally means that opensips (the core) has failed to send out the
packet over TCP. This will never auto-heal. A new TCP connection might
be set up, with a new subscription, but the odds of it being on the same
source port are next to none.

Instead, the 477s will cause lots of delays when the presence server
sends out NOTIFYs in batches, effectively blocking all opensips core
children while they're trying to connect to a closed/filtered tcp port.