GN: Unbreak exec_msg() broken in 1ccdca46d.
By moving from popen()+pclose() to custom_popen()+waitpid(), the stdin
of the subprocess was never closed. Causing the subprocess to wait until
infinity. This was broken between opensips 1.10.x and 1.11.x.

In master, the exec_msg() has been deprecated in favor of exec(). The
new exec() does pclose() the input stream. Although that looks like it
should be a fclose(). Seeing that popen3 does an fdopen(), so it won't
add the same properties that popen() would. (And besides, we don't
*want* to waitpid either at this point.)

NOTE: exec_str and exec_avp and friends may still be broken!