spiral: Fixes so upstream hop-by-hop ACKs don't get mangled
Locally generated ACKs don't turn up in spiral_pre_raw(), so they'd have no
direction cookies.

This changeset adds cookies to the INVITE R-URI. This is stored in the
transaction layer, so the tm-generated local hop-by-hop ACKs get that R-URI
cookie, which we can then read to determine the direction.

Fixes the problem of:

    A ---INVITE--> O
                   O ---INVITE'--> B
                   O <-----200'--- B
    A <-----200--- O
    A ------ACK--> O
                   O ------ACK'--> B

The above end-to-end ACK is fine, because it is not locally generated.

But now, watch the hop-by-hop ACK in case of an upstream INVITE:

                   O <--INVITE'--- B
    A <--INVITE--- O
    A ------418--> O
    A <-----ACK--- O
                   O ------418'--> B
                   O <-----ACK'--- B

Before this patch, that ACK to A would have a mangled call-id (ACK'),
because it defaulted to "downstream".

(cherry-picked from 6c0a608ce61e86fa083aaebbecadc11f6725e2bf)