GN: Do not expire pua PUBLISH before real end
The presence server can expire the record at its own leisurely pace. We
don't need to be exactly on time.

What we _don't_ want, is to be too early. When we do, in conjunction with
pua_usrloc, we would place devices in "closed" state for (at most)
update_period: the device ping-pongs between "open" and "closed".

See: opensips/opensips#3125